Move TaskHandler to own module
